+++ routes/index.js
@@ -1,6 +1,6 @@
 module.exports = {
-  'get /github/': 'index',
   'get /github': 'index',
+  'get /github/rate_limit': 'rate_limit',
   'get /github/:user': 'notFound',
   'get /github/:user/:repo': 'getRepo',
   'get /github/:user/:repo/FeatureServer': 'featureservice',
@@ -12,7 +12,6 @@ module.exports = {
   'get /github/:user/:repo/:file/FeatureServer': 'featureservice',
   'get /github/:user/:repo/:file/FeatureServer/:layer': 'featureservice',
   'get /github/:user/:repo/:file/FeatureServer/:layer/:method': 'featureservice',
-  'get /github/:user/:repo/:file/thumbnail': 'thumbnail',
   'get /github/:user/:repo/:file/tiles/:z/:x/:y.:format': 'tiles',
   'get /github/:user/:repo/:file/:layer/tiles/:z/:x/:y.:format': 'tiles',
   'get /github/:user/:repo/:file/drop': 'drop'