+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+* **controller**: preview route dependencies work on non-root mountpaths
+* **model**: local cache works (resolved issue from koop upstream)
+* **model**: github access tokens now work as expected
+
+### Changed
+* **provider**: name is now `github` instead of `Github`
+* **provider**: `status.version` moved to `version`
+* **controller**: removed `ctrl.Error`, using `koop-provider`'s `ctrl.errorResponse` method
+* **model**: simplify `find` and `drop` methods (use options object)
+
+### Added
+* **model**: looks for `KOOP_GITHUB_TOKEN` environmental variable if `config.ghtoken` isn't specified
+* **controller/routes**: added `rate_limit` route for checking github rate limit status
+* **controller**: support for JSONP callbacks ([`res.jsonp`](http://expressjs.com/api.html#res.jsonp))
+
+### Removed
+* **controller/routes**: thumbnail (Thumbnail generation no longer supported)
+
 ## [1.0.1] - 2015-08-13
 
 ### Changed